Add missing folder name suffix in bootloader paths
The paths of the non-autobaud bootloader binaries include an upload baud rate folder. These folder names have a "_baud"
suffix.

Previously, the `bootloader.file` properties for these bootloaders only had the integer component in the path. This
caused the "Burn Bootloader" operation to fail when configured to use a non-autobaud bootloader:

> avrdude ioerror() OS error: file ~/Library/Arduino15/packages/MegaCore/hardware/avr/3.0.1/bootloaders/urboot/atmega2560/watchdog_1_s/external_oscillator/16000000_hz/115200/uart3_rxj0_txj1/led+b7/urboot_atmega2560_pr_ee_ce.hex is not readable: No such file or directory